requirements-score.txt
pre-commit
pytest
pytest-mock
pandas
seaborn
pyyaml
